Move Chunkable from hashing to casper-node.
