+++ node/src/types.rs
@@ -4,6 +4,7 @@ pub(crate) mod appendable_block;
 mod available_block_range;
 mod block;
 pub mod chainspec;
+mod chunkable;
 mod deploy;
 pub mod error;
 mod exit_code;
@@ -33,6 +34,7 @@ pub(crate) use block::{
 };
 pub use chainspec::Chainspec;
 pub(crate) use chainspec::{ActivationPoint, ChainspecRawBytes};
+pub use chunkable::Chunkable;
 pub use datasize::DataSize;
 pub use deploy::{
     Approval, Deploy, DeployConfigurationFailure, DeployHash, DeployHeader, DeployMetadata,
